# Conflicts:
#	src/main/java/com/hanhwa_tae/gulhan/packages/query/service/PackageQueryService.java

+++ src/main/java/com/hanhwa_tae/gulhan/dashboard/query/mapper/DashboardQueryMapper.java
@@ -19,21 +19,21 @@ public interface DashboardQueryMapper {
 
     /* 패키지 판매 수량 및 수익 */
     List<PackageQuantityDTO> selectPackageList(DashboardPackageSearchRequest request);
-    int countPackage(DashboardPackageSearchRequest request);
+    long countPackage(DashboardPackageSearchRequest request);
 
     /* 기념품 판매 수량 및 수익 */
     List<GoodsQuantityDTO> selectGoodsList(DashboardGoodsSearchRequest request);
-    int countGoods(DashboardGoodsSearchRequest request);
+    long countGoods(DashboardGoodsSearchRequest request);
 
     /* 월별 기념품 & 패키지 수익 */
     List<MonthlyRevenueDTO> selectMonthlyList(DashboardMonthlySearchRequest request);
-    int countMonthly(DashboardMonthlySearchRequest request);
+    long countMonthly(DashboardMonthlySearchRequest request);
 
     /* 일별 기념품 & 패키지 수익 */
     List<DailyRevenueDTO> selectDailyList(DashboardDailySearchRequest request);
-    int countDaily(DashboardDailySearchRequest request);
+    long countDaily(DashboardDailySearchRequest request);
 
     /* 전체 주문내역 조회 - 패키지만 산 사람은 제외 */
     List<OrderAllListDTO> selectOrderList(DashboardOrderSearchRequest request);
-    int countOrders(DashboardOrderSearchRequest request);
+    long countOrders(DashboardOrderSearchRequest request);
 }
